* 'website-update' of github.com:zeo1995/pulsar: (432 commits)
  Fixed ordering issue in KeyShared dispatcher when adding consumer (#7106)
  Fix Duplicated messages are sent to dead letter topic #6960 (#7021)
  [Issue 2793][Doc]--Update the TLS hostname verification for CPP and Python clients (#7162)
  [Doc]--set netty mex frame size (#7174)
  [Doc] Update for the maximum message size (#7171)
  Fixed KeyShared consumers getting stuck on delivery (#7105)
  [#6003][pulsar-functions] Possibility to add builtin Functions (#6895)
  [Issue 6921][pulsar-broker-common] Replaced "Paths.get(...).getParent()", because it's system dependent and uses '\' as path separator on Windows (#6992)
  Improve broker unit test CI (#7173)
  Fix typo in exception message (#7027)
  Support KeyValue Schema Use Null Key And Null Value (#7139)
  [Doc]--Update documents for support consumer priority level in failover mode (#7136)
  Add schema config to cpp and cgo docs. (#7137)
  [Doc]--Update for the maximum message size (#7160)
  [C++] Expose ZSTD and Snappy compression to C API (#7014)
  [pulsar-proxy] add proxyLogLevel into config file (#6948)
  Add multi-hosts example for bookkeeperMetadataServiceUri (#6998)
  support for termination of partitioned topic (#6126)
  Use pure-java Air-Compressor instead of JNI based libraries (#5390)
  [Issues 5709]remove the namespace checking (#5716)
  ...

# Conflicts:
#	site2/website/scripts/split-swagger-by-version.js

+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/impl/EntryImpl.java
@@ -114,7 +114,7 @@ public final class EntryImpl extends AbstractCASReferenceCounted implements Entr
 
     @Override
     public byte[] getData() {
-        byte[] array = new byte[(int) data.readableBytes()];
+        byte[] array = new byte[data.readableBytes()];
         data.getBytes(data.readerIndex(), array);
         return array;
     }
